refactor(server): split router.go into cohesive files
The 923-line router.go fused four unrelated concerns. Split it into
same-package files, moving each declaration verbatim:

- router.go        route table, StartRouter, corsConfig, prometheusHandler
- env.go           Env struct, NewEnv, Shutdown, StartLockdownWatcher
- handlers.go      task/state/config/version handlers and auth glue
- handlers_lock.go deploy-lock handlers
- staticfs.go      safeFileSystem path-traversal guard and static serving
- websocket.go     connection registry, wsResponseWriter, WS handlers

No behavior change: identical declaration set (functions, types, and the
split constants keep their values), swagger annotations intact, imports
redistributed per file. Two pre-existing issues (the wsResponseWriter
hijack/shutdown header race and the release-path WebSocket notify) are
carried over unchanged; fixing them is separate work.
