+++ internal/server/router.go
@@ -1,214 +1,20 @@
 package server
 
 import (
-	"bufio"
-	"context"
-	"errors"
 	"fmt"
-	"io"
 	"log/slog"
-	"net"
 	"net/http"
 	"os"
 	"path/filepath"
-	"strconv"
 	"strings"
-	"sync"
 	"time"
 
-	"github.com/coder/websocket"
 	"github.com/gin-contrib/cors"
 	"github.com/gin-gonic/gin"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
-
-	"github.com/shini4i/argo-watcher/internal/argocd"
-	"github.com/shini4i/argo-watcher/internal/auth"
-	"github.com/shini4i/argo-watcher/internal/config"
-	"github.com/shini4i/argo-watcher/internal/models"
-	"github.com/shini4i/argo-watcher/internal/prometheus"
 )
 
-var version = "local"
-
-// maxTaskListLimit caps the page size accepted by GET /api/v1/tasks. The
-// underlying backends treat limit <= 0 as "no LIMIT clause", which would let
-// any caller drain the entire task table in a single request. The cap is
-// applied at the HTTP boundary so the data layer stays simple.
-const maxTaskListLimit = 1000
-
-var (
-	connectionsMutex sync.RWMutex
-	connections      []*websocket.Conn
-	closedConns      = make(map[*websocket.Conn]bool) // Track closed connections to prevent use-after-close
-)
-
-const (
-	deployLockEndpoint  = "/deploy-lock"
-	unauthorizedMessage = "You are not authorized to perform this action"
-	keycloakHeader      = "Keycloak-Authorization"
-	shutdownTimeout     = 10 * time.Second // Maximum time to wait for WebSocket goroutines during shutdown
-)
-
-// safeFileSystem wraps http.Dir with additional symlink protection.
-// This provides defense-in-depth beyond http.Dir's built-in path sanitization.
-type safeFileSystem struct {
-	root     http.Dir
-	basePath string
-}
-
-// validatePath checks if a path is safe before any I/O operations.
-// Returns the cleaned path if valid, or an error if the path would escape the base directory.
-// This performs validation without any I/O operations by checking the cleaned path.
-func (fs safeFileSystem) validatePath(name string) (string, error) {
-	// Clean the path to remove any .. or . components
-	cleanName := filepath.Clean("/" + name)
-
-	// Construct the would-be full path and verify it stays within bounds
-	// filepath.Join handles path separators and cleaning
-	fullPath := filepath.Join(fs.basePath, cleanName)
-
-	// Clean the full path to resolve any remaining . or .. components
-	cleanedFull := filepath.Clean(fullPath)
-
-	// Verify the cleaned path is still within the base directory
-	// Check for exact match (root directory) or proper prefix with separator
-	if cleanedFull != fs.basePath && !strings.HasPrefix(cleanedFull, fs.basePath+string(filepath.Separator)) {
-		slog.Debug("blocked path traversal attempt",
-			"requested_path", name,
-			"resolved_path", cleanedFull,
-			"base_path", fs.basePath)
-		return "", os.ErrPermission
-	}
-
-	return cleanName, nil
-}
-
-// Open implements http.FileSystem interface with path validation and symlink protection.
-func (fs safeFileSystem) Open(name string) (http.File, error) {
-	// Validate path before any I/O operation
-	cleanName, err := fs.validatePath(name)
-	if err != nil {
-		return nil, err
-	}
-
-	// Open using the validated clean path
-	// http.Dir.Open has its own path sanitization as additional protection
-	f, err := fs.root.Open(cleanName)
-	if err != nil {
-		return nil, err
-	}
-
-	// Additional symlink protection: verify the real path is within bounds
-	osFile, ok := f.(*os.File)
-	if !ok {
-		return f, nil
-	}
-
-	realPath, err := filepath.EvalSymlinks(osFile.Name())
-	if err != nil {
-		_ = f.Close() // #nosec G104 - best effort cleanup
-		return nil, err
-	}
-
-	if !strings.HasPrefix(realPath, fs.basePath+string(filepath.Separator)) && realPath != fs.basePath {
-		slog.Debug("blocked symlink escape attempt",
-			"requested_path", name,
-			"resolved_path", realPath,
-			"base_path", fs.basePath)
-		_ = f.Close() // #nosec G104 - best effort cleanup
-		return nil, os.ErrPermission
-	}
-
-	return f, nil
-}
-
-// wsResponseWriter wraps gin's ResponseWriter to provide proper WebSocket hijacking.
-// gin's ResponseWriter fails Hijack() after WriteHeader() is called, but WebSocket
-// upgrade requires both operations. This wrapper hijacks the connection early
-// (before WriteHeader) and stores the raw connection for later use.
-type wsResponseWriter struct {
-	gin.ResponseWriter
-	conn          net.Conn
-	brw           *bufio.ReadWriter
-	headerWritten bool
-}
-
-// Hijack returns the pre-hijacked connection, bypassing gin's "already written" check.
-func (w *wsResponseWriter) Hijack() (net.Conn, *bufio.ReadWriter, error) {
-	if w.conn == nil {
-		return nil, nil, errors.New("connection was not pre-hijacked")
-	}
-	return w.conn, w.brw, nil
-}
-
-// Write writes data through the buffered writer to maintain consistency.
-func (w *wsResponseWriter) Write(data []byte) (int, error) {
-	if w.brw == nil {
-		return 0, errors.New("buffered writer not available")
-	}
-	n, err := w.brw.Write(data)
-	if err != nil {
-		return n, err
-	}
-	return n, w.brw.Flush()
-}
-
-// WriteHeader writes the status line and headers through the buffered writer.
-// Note: The http.ResponseWriter interface does not allow WriteHeader to return an error,
-// so errors are logged but cannot be propagated to the caller.
-// Per http.ResponseWriter contract, multiple calls should be no-ops after the first.
-func (w *wsResponseWriter) WriteHeader(code int) {
-	if w.headerWritten {
-		return
-	}
-	w.headerWritten = true
-
-	if w.brw == nil {
-		slog.Error("buffered writer not available during WriteHeader")
-		return
-	}
-	statusLine := fmt.Sprintf("HTTP/1.1 %d %s\r\n", code, http.StatusText(code))
-	if _, err := w.brw.WriteString(statusLine); err != nil {
-		slog.Error("failed to write status line during WebSocket upgrade", "error", err)
-		return
-	}
-	if err := w.Header().Write(w.brw); err != nil {
-		slog.Error("failed to write headers during WebSocket upgrade", "error", err)
-		return
-	}
-	if _, err := w.brw.WriteString("\r\n"); err != nil {
-		slog.Error("failed to write header terminator during WebSocket upgrade", "error", err)
-		return
-	}
-	if err := w.brw.Flush(); err != nil {
-		slog.Error("failed to flush WebSocket upgrade response", "error", err)
-	}
-}
-
-// Env reference: https://www.alexedwards.net/blog/organising-database-access
-type Env struct {
-	// environment configurations
-	config *config.ServerConfig
-	// argo argo
-	argo *argocd.Argo
-	// argo updater
-	updater *argocd.ArgoStatusUpdater
-	// metrics
-	metrics *prometheus.Metrics
-	// deploy lock
-	lockdown *Lockdown
-	// enabled auth strategies
-	strategies map[string]auth.AuthStrategy
-	// authenticator orchestrates registered strategies
-	authenticator *auth.Authenticator
-	// shutdownCh signals graceful shutdown to all WebSocket goroutines.
-	// Using a channel instead of storing context.Context follows Go best practices.
-	shutdownCh chan struct{}
-	// shutdownOnce ensures Shutdown() can be called multiple times safely.
-	shutdownOnce sync.Once
-	// connWg tracks active WebSocket connection goroutines for graceful shutdown.
-	connWg sync.WaitGroup
-}
+const deployLockEndpoint = "/deploy-lock"
 
 // CreateRouter initialize router.
 func (env *Env) CreateRouter() *gin.Engine {
@@ -316,40 +122,6 @@ func (env *Env) CreateRouter() *gin.Engine {
 	return router
 }
 
-// createStaticFileHandler returns a handler for serving static files with SPA fallback.
-// It attempts to serve the requested file, falling back to index.html for SPA routing.
-func (env *Env) createStaticFileHandler(fs safeFileSystem, staticPath string) gin.HandlerFunc {
-	return func(c *gin.Context) {
-		if tryServeStaticFile(c, fs) {
-			return
-		}
-		// Fall back to index.html for SPA routing
-		c.File(filepath.Join(staticPath, "index.html"))
-	}
-}
-
-// tryServeStaticFile attempts to serve a static file and returns true if successful.
-func tryServeStaticFile(c *gin.Context, fs safeFileSystem) bool {
-	f, err := fs.Open(c.Request.URL.Path)
-	if err != nil {
-		return false
-	}
-	defer f.Close()
-
-	stat, err := f.Stat()
-	if err != nil || stat.IsDir() {
-		return false
-	}
-
-	rs, ok := f.(io.ReadSeeker)
-	if !ok {
-		return false
-	}
-
-	http.ServeContent(c.Writer, c.Request, stat.Name(), stat.ModTime(), rs)
-	return true
-}
-
 // StartRouter creates and returns an HTTP server configured with the given router.
 // The caller is responsible for starting the server and handling graceful shutdown.
 func (env *Env) StartRouter(router *gin.Engine) *http.Server {
@@ -359,56 +131,6 @@ func (env *Env) StartRouter(router *gin.Engine) *http.Server {
 		Addr:              routerBind,
 		Handler:           router,
 		ReadHeaderTimeout: 10 * time.Second, // Prevent Slowloris attacks
-	}
-}
-
-// lockdownPollInterval is how often the scheduled-lockdown watcher re-evaluates
-// the lock state to detect schedule boundary transitions. Schedules have
-// minute granularity, so a one-minute tick bounds the notification lag.
-const lockdownPollInterval = time.Minute
-
-// StartLockdownWatcher launches a background goroutine that notifies WebSocket
-// clients when a scheduled lockdown window automatically begins or ends. It is a
-// no-op when no schedules are configured, since manual set/release already
-// notify clients directly. The goroutine is tracked by connWg and stops when the
-// shutdown channel is closed.
-func (env *Env) StartLockdownWatcher() {
-	if len(env.lockdown.Schedules) == 0 {
-		return
-	}
-
-	env.connWg.Add(1)
-	go func() {
-		defer env.connWg.Done()
-		env.lockdown.WatchTransitions(env.shutdownCh, lockdownPollInterval, notifyWebSocketClients)
-	}()
-}
-
-// Shutdown gracefully shuts down the server and all WebSocket connections.
-// This method is safe to call multiple times. It blocks until all WebSocket
-// goroutines have finished or the shutdown timeout is reached. If the timeout
-// is reached, some goroutines may still be running but should exit shortly as
-// they observe the closed shutdownCh. Any long-running WebSocket writes are
-// bounded by their own 5-second timeout in checkConnection.
-func (env *Env) Shutdown() {
-	if env.shutdownCh != nil {
-		env.shutdownOnce.Do(func() {
-			close(env.shutdownCh)
-		})
-	}
-
-	// Wait for all WebSocket connection goroutines to finish with timeout
-	done := make(chan struct{})
-	go func() {
-		env.connWg.Wait()
-		close(done)
-	}()
-
-	select {
-	case <-done:
-		slog.Debug("All WebSocket connections closed gracefully")
-	case <-time.After(shutdownTimeout):
-		slog.Warn("Shutdown timeout reached, some WebSocket goroutines may still be running")
 	}
 }
 
@@ -446,478 +168,4 @@ func prometheusHandler() gin.HandlerFunc {
 	return func(c *gin.Context) {
 		ph.ServeHTTP(c.Writer, c.Request)
 	}
-}
-
-// getVersion godoc
-// @Summary Get the version of the server
-// @Description Get the version of the server
-// @Tags frontend
-// @Success 200 {string} string
-// @Router /api/v1/version [get]
-func (env *Env) getVersion(c *gin.Context) {
-	c.JSON(http.StatusOK, version)
-}
-
-// addTask godoc
-// @Summary Add a new task
-// @Description Add a new task
-// @Tags backend
-// @Accept json
-// @Produce json
-// @Param task body models.Task true "Task"
-// @Success 202 {object} models.TaskStatus
-// @Failure 401 {object} models.TaskStatus
-// @Failure 406 {object} models.TaskStatus
-// @Router /api/v1/tasks [post]
-func (env *Env) addTask(c *gin.Context) {
-	var task models.Task
-
-	err := c.ShouldBindJSON(&task)
-	if err != nil {
-		slog.Error(fmt.Sprintf("couldn't process new task, got the following error: %s", err))
-		c.JSON(http.StatusNotAcceptable, models.TaskStatus{
-			Status: "invalid payload",
-			Error:  err.Error(),
-		})
-		return
-	}
-
-	// we need to handle cases when deploy lock is set either manually or by cron
-	if env.lockdown.IsLocked() {
-		slog.Warn("deploy lock is set, rejecting the task")
-		c.JSON(http.StatusNotAcceptable, models.TaskStatus{
-			Status: "rejected",
-			Error:  "lockdown is active, deployments are not accepted",
-		})
-		return
-	}
-
-	tokenValid, err := env.validateToken(c, "")
-	if err != nil {
-		// A non-nil error means the strategy was invoked and rejected the
-		// token: a client mistake, not a server failure. Return 401 with
-		// the reason so the client can show something actionable.
-		slog.Warn(fmt.Sprintf("rejecting task: %s", err))
-		c.JSON(http.StatusUnauthorized, models.TaskStatus{
-			Status: unauthorizedMessage,
-			Error:  err.Error(),
-		})
-		return
-	}
-
-	task.Validated = tokenValid
-
-	newTask, err := env.argo.AddTask(task)
-	if err != nil {
-		slog.Error(fmt.Sprintf("Couldn't process new task. Got the following error: %s", err))
-		c.JSON(http.StatusServiceUnavailable, models.TaskStatus{
-			Status: "down",
-			Error:  err.Error(),
-		})
-		return
-	}
-
-	// start rollout monitor
-	go env.updater.WaitForRollout(*newTask)
-
-	// return information about created task
-	c.JSON(http.StatusAccepted, models.TaskStatus{
-		Id:     newTask.Id,
-		Status: models.StatusAccepted,
-	})
-}
-
-// getState godoc
-// @Summary Get state content
-// @Description Get all tasks that match the provided parameters
-// @Tags backend, frontend
-// @Param app query string false "App name"
-// @Param status query string false "Task status (e.g. 'in progress', 'failed', 'deployed', 'cancelled')"
-// @Param from_timestamp query int true "From timestamp" default(1648390029)
-// @Param to_timestamp query int false "To timestamp"
-// @Param limit query int false "Maximum number of tasks to return (1-1000, defaults to 1000)"
-// @Param offset query int false "Number of tasks to skip before returning results"
-// @Success 200 {object} models.TasksResponse
-// @Router /api/v1/tasks [get]
-func (env *Env) getState(c *gin.Context) {
-	startTime, err := strconv.ParseFloat(c.Query("from_timestamp"), 64)
-	if err != nil && c.Query("from_timestamp") != "" {
-		slog.Debug("invalid from_timestamp, defaulting to 0", "from_timestamp", c.Query("from_timestamp"))
-	}
-	endTime, err := strconv.ParseFloat(c.Query("to_timestamp"), 64)
-	if err != nil && c.Query("to_timestamp") != "" {
-		slog.Debug("invalid to_timestamp, defaulting to current time", "to_timestamp", c.Query("to_timestamp"))
-	}
-	if endTime == 0 {
-		endTime = float64(time.Now().Unix())
-	}
-	app := c.Query("app")
-	status := c.Query("status")
-	if status != "" && !models.IsAllowedTaskStatus(status) {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "unsupported status filter"})
-		return
-	}
-
-	limit, err := strconv.Atoi(c.Query("limit"))
-	if err != nil && c.Query("limit") != "" {
-		slog.Debug("invalid limit, defaulting to 0", "limit", c.Query("limit"))
-	}
-	offset, err := strconv.Atoi(c.Query("offset"))
-	if err != nil && c.Query("offset") != "" {
-		slog.Debug("invalid offset, defaulting to 0", "offset", c.Query("offset"))
-	}
-	if limit <= 0 || limit > maxTaskListLimit {
-		limit = maxTaskListLimit
-	}
-	if offset < 0 {
-		offset = 0
-	}
-
-	c.JSON(http.StatusOK, env.argo.GetTasks(startTime, endTime, app, status, limit, offset))
-}
-
-// getTaskStatus godoc
-// @Summary Get the status of a task
-// @Description Get the status of a task
-// @Param id path string true "Task id" default(9185fae0-add5-11ec-87f3-56b185c552fa)
-// @Tags backend
-// @Produce json
-// @Success 200 {object} models.TaskStatus
-// @Failure 404 {object} models.TaskStatus
-// @Router /api/v1/tasks/{id} [get]
-func (env *Env) getTaskStatus(c *gin.Context) {
-	id := c.Param("id")
-	task, err := env.argo.State.GetTask(id)
-
-	if err != nil {
-		c.JSON(http.StatusNotFound, models.TaskStatus{
-			Id:    id,
-			Error: err.Error(),
-		})
-	} else {
-		c.JSON(http.StatusOK, models.TaskStatus{
-			Id:           task.Id,
-			Created:      task.Created,
-			Updated:      task.Updated,
-			App:          task.App,
-			Author:       task.Author,
-			Project:      task.Project,
-			Images:       task.Images,
-			Status:       task.Status,
-			StatusReason: task.StatusReason,
-		})
-	}
-}
-
-// healthz godoc
-// @Summary Check if the server is healthy
-// @Description Check if the argo-watcher is ready to process new tasks
-// @Tags service
-// @Produce json
-// @Success 200 {object} models.HealthStatus
-// @Failure 503 {object} models.HealthStatus
-// @Router /healthz [get]
-func (env *Env) healthz(c *gin.Context) {
-	if env.argo.SimpleHealthCheck() {
-		c.JSON(http.StatusOK, models.HealthStatus{
-			Status: "up",
-		})
-	} else {
-		c.JSON(http.StatusServiceUnavailable, models.HealthStatus{
-			Status: "down",
-		})
-	}
-
-}
-
-// getConfig godoc
-// @Summary Get the configuration of the server (excluding sensitive data)
-// @Description Get the configuration of the server (excluding sensitive data)
-// @Tags backend
-// @Produce json
-// @Success 200 {object} config.ServerConfig
-// @Router /api/v1/config [get]
-func (env *Env) getConfig(c *gin.Context) {
-	c.JSON(http.StatusOK, env.config)
-}
-
-// requireKeycloakAuth validates the Keycloak token when Keycloak is enabled.
-// It returns true if validation passes (or Keycloak is disabled). On failure
-// the response distinguishes:
-//   - 401 with "authentication required" when no auth header was sent.
-//   - 401 with the strategy's reason when the token was rejected.
-//
-// Strategy transport/parse failures (e.g. Keycloak unreachable) are also
-// returned as 401 with a sanitized "token validation failed" message; full
-// details land in the server log only.
-func (env *Env) requireKeycloakAuth(c *gin.Context) bool {
-	if !env.config.Keycloak.Enabled {
-		return true
-	}
-
-	valid, err := env.validateToken(c, keycloakHeader)
-	if valid {
-		return true
-	}
-
-	if err != nil {
-		// Strategy was invoked and rejected the token. Surface the reason.
-		slog.Warn(fmt.Sprintf("User tried %s %s with invalid token: %s",
-			c.Request.Method, c.Request.URL, err))
-		c.JSON(http.StatusUnauthorized, models.TaskStatus{
-			Status: unauthorizedMessage,
-			Error:  err.Error(),
-		})
-		return false
-	}
-
-	// (false, nil): no auth header sent at all.
-	slog.Warn(fmt.Sprintf("User tried %s %s without authentication", c.Request.Method, c.Request.URL))
-	c.JSON(http.StatusUnauthorized, models.TaskStatus{
-		Status: unauthorizedMessage,
-		Error:  "authentication required (set " + keycloakHeader + " header)",
-	})
-	return false
-}
-
-// SetDeployLock godoc
-// @Summary Set deploy lock
-// @Description Set deploy lock. Only available when Keycloak is enabled; requires a valid Keycloak session.
-// @Tags frontend
-// @Success 200 {string} string
-// @Failure 401 {object} models.TaskStatus
-// @Router /api/v1/deploy-lock [post]
-func (env *Env) SetDeployLock(c *gin.Context) {
-	if !env.requireKeycloakAuth(c) {
-		return
-	}
-
-	env.lockdown.SetLock()
-
-	slog.Debug("deploy lock is set")
-
-	notifyWebSocketClients("locked")
-
-	c.JSON(http.StatusOK, "deploy lock is set")
-}
-
-// ReleaseDeployLock godoc
-// @Summary Release deploy lock
-// @Description Release deploy lock. Only available when Keycloak is enabled; requires a valid Keycloak session.
-// @Tags frontend
-// @Success 200 {string} string
-// @Failure 401 {object} models.TaskStatus
-// @Router /api/v1/deploy-lock [delete]
-func (env *Env) ReleaseDeployLock(c *gin.Context) {
-	if !env.requireKeycloakAuth(c) {
-		return
-	}
-
-	env.lockdown.ReleaseLock()
-
-	slog.Debug("deploy lock is released")
-
-	notifyWebSocketClients("unlocked")
-
-	c.JSON(http.StatusOK, "deploy lock is released")
-}
-
-// isDeployLockSet godoc
-// @Summary Check if deploy lock is set
-// @Description Check if deploy lock is set
-// @Tags frontend
-// @Success 200 {boolean} boolean
-// @Router /api/v1/deploy-lock [get]
-func (env *Env) isDeployLockSet(c *gin.Context) {
-	c.JSON(http.StatusOK, env.lockdown.IsLocked())
-}
-
-// validateToken validates the incoming request using the configured authentication strategies.
-// When allowedAuthStrategy is empty, the validation delegates to the default authenticator,
-// which returns the last validation error when no strategies succeed. When allowedAuthStrategy
-// is provided, validation is restricted to that specific strategy header via the authenticator.
-func (env *Env) validateToken(c *gin.Context, allowedAuthStrategy string) (bool, error) {
-	if allowedAuthStrategy == "" {
-		return env.authenticator.Validate(c.Request)
-	}
-
-	return env.authenticator.ValidateStrategy(c.Request, allowedAuthStrategy)
-}
-
-// handleWebSocketConnection accepts a WebSocket connection, adds it to a slice,
-// and initiates a goroutine to ping the connection regularly. If WebSocket
-// acceptance fails, an error is logged. The goroutine serves to monitor
-// the connection's activity and removes it from the slice if it's inactive.
-func (env *Env) handleWebSocketConnection(c *gin.Context) {
-	options := &websocket.AcceptOptions{
-		InsecureSkipVerify: env.config.DevEnvironment, // It will disable websocket host validation if set to true
-	}
-
-	// Pre-hijack the connection BEFORE WriteHeader is called
-	// gin's ResponseWriter fails Hijack after WriteHeader, so we hijack first
-	hijacker, ok := c.Writer.(http.Hijacker)
-	if !ok {
-		slog.Error("ResponseWriter does not support hijacking")
-		c.String(http.StatusInternalServerError, "WebSocket not supported")
-		return
-	}
-
-	netConn, brw, err := hijacker.Hijack()
-	if err != nil {
-		slog.Error("failed to hijack connection for WebSocket", "error", err)
-		// After a failed hijack, the connection state is unknown and we cannot reliably
-		// write a response. The client connection will eventually timeout.
-		return
-	}
-
-	// Create wrapper with pre-hijacked connection
-	wrappedWriter := &wsResponseWriter{
-		ResponseWriter: c.Writer,
-		conn:           netConn,
-		brw:            brw,
-	}
-
-	conn, err := websocket.Accept(wrappedWriter, c.Request, options)
-	if err != nil {
-		slog.Error("failed to accept websocket connection", "error", err)
-		_ = netConn.Close() // #nosec G104 - best effort cleanup, already in error path
-		return
-	}
-
-	// Append the connection before starting the goroutine
-	connectionsMutex.Lock()
-	connections = append(connections, conn)
-	connectionsMutex.Unlock()
-
-	// Track the goroutine for graceful shutdown
-	env.connWg.Add(1)
-	go env.checkConnection(conn)
-}
-
-// checkConnection is a method for the Env struct that continuously checks the
-// health of a WebSocket connection by sending periodic "heartbeat" messages.
-func (env *Env) checkConnection(c *websocket.Conn) {
-	defer env.connWg.Done()
-
-	ticker := time.NewTicker(time.Second * 30)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-env.shutdownCh:
-			_ = c.Close(websocket.StatusGoingAway, "server shutdown")
-			removeWebSocketConnection(c)
-			return
-		case <-ticker.C:
-			// we are not using c.Ping here, because it's not working as expected
-			// for some reason it's failing even if the connection is still alive
-			// if you know how to fix it, please open an issue or PR
-			ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-			if c.Write(ctx, websocket.MessageText, []byte("heartbeat")) != nil {
-				cancel()
-				_ = c.Close(websocket.StatusNormalClosure, "heartbeat failed")
-				removeWebSocketConnection(c)
-				return
-			}
-			cancel()
-		}
-	}
-}
-
-// notifyWebSocketClients is a function that sends a message to all active WebSocket clients.
-// It iterates over the global connections slice, which contains all active WebSocket connections,
-// and sends the provided message to each connection using the wsjson.Write function.
-// If an error occurs while sending the message to a connection (for example, if the connection has been closed),
-// it removes the connection from the connections slice to prevent further attempts to send messages to it.
-func notifyWebSocketClients(message string) {
-	var wg sync.WaitGroup
-
-	// Copy connections slice under mutex to avoid race condition during iteration
-	// Also filter out closed connections
-	connectionsMutex.RLock()
-	connsCopy := make([]*websocket.Conn, 0, len(connections))
-	for _, c := range connections {
-		if !closedConns[c] {
-			connsCopy = append(connsCopy, c)
-		}
-	}
-	connectionsMutex.RUnlock()
-
-	for _, conn := range connsCopy {
-		wg.Add(1)
-
-		go func(c *websocket.Conn, message string) {
-			defer wg.Done()
-			ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-			defer cancel()
-			if c.Write(ctx, websocket.MessageText, []byte(message)) != nil {
-				_ = c.Close(websocket.StatusNormalClosure, "write failed")
-				removeWebSocketConnection(c)
-			}
-		}(conn, message)
-	}
-
-	wg.Wait()
-}
-
-// removeWebSocketConnection is a helper function that removes a WebSocket connection
-// from the global connections slice. It is used to clean up connections that are no longer active.
-// The function takes a WebSocket connection as an argument and removes it from the connections slice.
-// It uses a mutex to prevent concurrent access to the connections slice, ensuring thread safety.
-// Note: Callers are responsible for closing the connection before calling this function.
-func removeWebSocketConnection(conn *websocket.Conn) {
-	connectionsMutex.Lock()
-	defer connectionsMutex.Unlock()
-
-	// Mark as closed first to prevent use-after-close during concurrent access
-	closedConns[conn] = true
-
-	for i := range connections {
-		if connections[i] == conn {
-			connections = append(connections[:i], connections[i+1:]...)
-			break
-		}
-	}
-
-	// Clean up closedConns entry to prevent memory leak
-	delete(closedConns, conn)
-}
-
-// NewEnv initializes a new Env instance.
-// This function is used to set up the environment for the application's main operation, including setting configurations, initializing Argo service, and metrics.
-func NewEnv(serverConfig *config.ServerConfig, argo *argocd.Argo, metrics *prometheus.Metrics, updater *argocd.ArgoStatusUpdater) (*Env, error) {
-	var env *Env
-	var err error
-
-	env = &Env{
-		config:     serverConfig,
-		argo:       argo,
-		metrics:    metrics,
-		updater:    updater,
-		shutdownCh: make(chan struct{}),
-	}
-
-	if env.lockdown, err = NewLockdown(serverConfig.LockdownSchedule); err != nil {
-		return nil, err
-	}
-
-	env.strategies = map[string]auth.AuthStrategy{
-		"ARGO_WATCHER_DEPLOY_TOKEN": auth.NewDeployTokenAuthService(env.config.DeployToken),
-	}
-
-	if env.config.Keycloak.Enabled {
-		keycloakService, keycloakErr := auth.NewKeycloakAuthService(env.config)
-		if keycloakErr != nil {
-			return nil, fmt.Errorf("failed to initialize keycloak auth: %w", keycloakErr)
-		}
-		env.strategies[keycloakHeader] = keycloakService
-	}
-
-	if env.config.JWTSecret != "" {
-		env.strategies["Authorization"] = auth.NewJWTAuthService(env.config.JWTSecret)
-	}
-
-	env.authenticator = auth.NewAuthenticator(env.strategies)
-
-	return env, nil
 }
